pyproject.toml
[project]
name = "tiny-transformer-lab"
version = "0.1.0"
description = "Tiny Transformer Lab — infra & data tools"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Andrei" }
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.6",
    "pyyaml>=6",
    "pandas>=2.2",
    "pyarrow>=17",
    "numpy>=1.26",
    "tqdm>=4.66",
    "mlflow>=2.16.0",
    "rich>=13",
    "loguru>=0.7"
]

[project.scripts]
ttlab = "ttlab.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "mypy>=1.10",
    "ruff>=0.6",
    "black>=24",
    "pip-audit>=2.7"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
